requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.6
numpy==1.26.4
packaging==25.0
pandas==2.2.1
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
